client: Store boolean in window for the RxPlayer
This commit adds a `__Rx_PLAYER_DEBUG_MODE__` boolean to `true` in
`window`, so the RxPlayer can then read it and put iself in a `"DEBUG"`
LogLevel (and perhaps other things in the future) when it is present.

This allows to simplify the usage of RxPaired to debug the RxPlayer in
various applications: it will become unnecessary to also have to raise
log verbosity manually, the only action will be to deploy the client
script on the device.