Server: Allow the creation of a bundle already containing dependencies
I am working currently with debugging people which are not well-versed
in JavaScript/Node.JS tools such as npm, npm scripts and so on.

Because we want them to rely on our tools without complexifying their
setup, I propose that we add the possibility to produce a bundle through
the `npm run bundle` script.

We can then transmit to them that bundle, that they will run like this:
```sh
node ./bundle.cjs --no-passwork -f # (-f == create log files)
```

Which when coupled with our client file deployed on the wanted device is all
they need to do to communicate us logs.

That bundle will already have our dependencies (beside node ones)
bundled, meaning that no `npm install` will be needed.

I relied on esbuild for the bundling part because of my familiarity with
its options. However, I noticed that producing the "ESM" bundle was
leading to a script which failed at runtime due to a "dynamic require"
performed by our "commander" dependency.
Because I didn't want to spend too much time on it as such issues often
are, I decided to produce a commonjs bundle.

The commonjs bundle did not have this issue but it had another simpler
to resolve one: top-level await is not authorized in those scripts and
we had one at initialization.

I chose to ""fix"" it by just making the `readFile` call ultimately creating
the need for that `await` as synchronous. This should not change
anything in the current code as it was top-level awaited at the only
place the corresponding function was called, though I guess this
solution could be not considered elegant.

Another possibility though could be to try with different bundling
options / different bundles / different dependencies but this seems much
more time consuming.

Thoughts?