Re-try to produce a server bundle with integrated dependencies
This is a re-attempt of #3 which had the default of having to be a
CommonJS bundle due to the commander library not being compatible with
esbuild way of producing esm outputs (and producing CommonJS outputs
doesn't work without modifications due to top-level await being
present).

This time, I chose to remove the commander dependency as we just
needed it to parse options and display an help screen.

I did it, and error messages when an argument is unknown or not in the
right format should be nicer in a way I suppose.

However, it still doesn't work because turns out that our only other
dependency "ws" (which performs our lower-level WebSockect connection
handling), also causes the same issue commander caused!

So, well this commit doesn't fix the issue, but it still has some value
which made me consider whether we would want to integrate this in some
form. We could also rely on this as a first step to truly fix the issue.